requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
openai==1.50.0
yt-dlp==2024.10.22
aiosqlite==0.20.0
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
Pillow==12.1.0
imagehash==4.3.2
stripe==14.3.0
sendgrid==6.12.5
psutil==7.2.2
